api/pyproject.toml
[tool.poetry]
name = "api"
version = "1.0.0a"
description = ""
authors = ["srusskih <smilemouse2002@ukr.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.9"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-django = "^4.7.0"
faker = "^22.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79
